edit application-dev.yml      ArchivesManageApplicationTests.java      CommTableFieldMapper.java      CommTableFieldMapper.xml      ICommTableFieldService.java      ICommTableFieldServiceImpl.java      SysArchivesLibraryField.java      SysArchivesLibraryFieldFormDto.java      SysArchivesLibraryFieldServiceImpl.java      SysArchivesLibraryServiceImpl.java      SysTemplateLibraryField.java      SysTemplateLibraryFieldFormDto.java files

+++ archives-manage-common/src/main/java/com/yintu/rixing/dto/system/SysTemplateLibraryFieldFormDto.java
@@ -36,11 +36,11 @@ public class SysTemplateLibraryFieldFormDto extends IdDto {
 
     @ApiModelProperty(value = "字段是否必填 1.是 0.否", required = true)
     @NotNull
-    private Integer required;
+    private Short required;
 
     @ApiModelProperty(value = "字段是否是索引 1.是 0.否", required = true)
     @NotNull
-    private Integer index;
+    private Short index;
 
     @ApiModelProperty(value = "字段顺序", required = true)
     @NotNull
