+++ archives-manage-entity/src/main/java/com/yintu/rixing/system/SysTemplateLibraryField.java
@@ -40,11 +40,11 @@ public class SysTemplateLibraryField extends BaseEntity {
 
     @ApiModelProperty(value = "字段是否必填 1.是 0.否", position = 9)
     @TableField("required")
-    private Integer required;
+    private Short required;
 
     @ApiModelProperty(value = "字段是否是索引 1.是 0.否", position = 10)
     @TableField("`index`")
-    private Integer index;
+    private Short index;
 
     @ApiModelProperty(value = "字段顺序", position = 11)
     @TableField("`order`")
